# docs/openapi.yaml
openapi: "3.0.3"
info:
  title: PulseScore API
  version: "1.0.0"
  description: |
    Customer health scoring platform API.
    PulseScore ingests data from integrations (Stripe, etc.), computes health scores,
    and surfaces risk analytics through a REST API consumed by the React dashboard.
servers:
  - url: /api/v1
    description: API v1
security:
  - bearerAuth: []

  # ── Tags ──
...
paths:
...
  /auth/password-reset/request:
    post:
      tags: [Auth]
      summary: Request a password reset email
      security: []
      operationId: requestPasswordReset
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email:
                  type: string
                  format: email
      responses:
        "200":
          description: Always returns success to prevent email enumeration
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /auth/password-reset/complete:
    post:
      tags: [Auth]
      summary: Complete password reset with token
      security: []
      operationId: completePasswordReset
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token, new_password]
              properties:
                token:
                  type: string
                new_password:
                  type: string
                  minLength: 8
      responses:
        "200":
          description: Password reset successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          $ref: "#/components/responses/ValidationError"

  # ── Invitations (public) ───────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
...
    Unauthorized:
      description: Authentication required or failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    ValidationError:
      description: Validation failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ValidationErrorResponse"
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
    ValidationErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
        field:
          type: string
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
...